fix(devtools):remove leading underscore from class name
