+++ devtools/projects/ng-devtools-backend/src/lib/utils.ts
@@ -6,6 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+import { InjectionToken, Type } from "@angular/core";
+
 export const runOutsideAngular = (f: () => any): void => {
   const w = window as any;
   if (!w.Zone || !w.Zone.current) {
@@ -34,3 +36,31 @@ export const isCustomElement = (node: Node) => {
   const tagName = node.tagName.toLowerCase();
   return !!customElements.get(tagName);
 };
+
+export const valueToLabel = (value: any): string => {
+  if (isInjectionToken(value)) {
+    return `InjectionToken(${value['_desc']})`;
+  }
+
+  if (typeof value === 'object') {
+    return stripUnderscore(value.constructor.name);
+  }
+
+  if (typeof value === 'function') {
+    return stripUnderscore(value.name);
+  }
+
+  return stripUnderscore(value);
+};
+
+export function stripUnderscore(str: string): string {
+  if (str.startsWith('_')) {
+    return str.slice(1);
+  }
+
+  return str;
+}
+
+export function isInjectionToken(token: Type<unknown>|InjectionToken<unknown>): boolean {
+  return token.constructor.name === 'InjectionToken';
+}
